+++ CHANGELOG.md
@@ -62,6 +62,7 @@ All notable changes to this project will be documented here. The format is based
 - Contact geometry can now be translated+rotated in the model editor tab's 3D viewport (#596)
 - Hotfixed a bug from OpenSim where inertia edits were not applied to the model (#597, related: opensim-core/#3395,
   thanks to @jesse-gilmer for spotting this :))
+- The spherical end-caps of muscle geometry now align better with the muscle cylinders they are attached to (#593)
 
 ## [0.3.2] - 2023/01/09